requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiokafka==0.12.0
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.0.5
async-timeout==5.0.1
asyncpg==0.31.0
attrs==25.4.0
babel==2.17.0
beautifulsoup4==4.14.3
black==23.12.1
bleach==6.3.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.7
colorama==0.4.6
comm==0.2.3
contourpy==1.3.2
coverage==7.12.0
cramjam==2.11.0
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
dnspython==2.8.0
executing==2.2.1
fastjsonschema==2.21.2
fastparquet==2023.10.1
flake8==7.0.0
fonttools==4.61.0
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.12.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
ipykernel==7.1.0
ipython==8.18.1
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.1
isoduration==20.11.0
isort==5.13.2
jedi==0.19.2
Jinja2==3.1.6
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.5.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
kafka-python==2.0.2
kiwisolver==1.4.9
lark==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.8.2
matplotlib-inline==0.2.1
mccabe==0.7.0
mdurl==0.1.2
mistune==3.1.4
multidict==6.7.0
mypy==1.8.0
mypy_extensions==1.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.5.0
notebook_shim==0.2.4
numpy==1.24.4
orjson==3.9.12
overrides==7.7.0
packaging==25.0
pandas==2.1.4
pandocfilters==1.5.1
parso==0.8.5
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.0
plotly==5.17.0
pluggy==1.6.0
postgres==4.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
psutil==7.1.3
psycopg2-binary==2.9.9
psycopg2-pool==1.2
pure_eval==0.2.3
pyarrow==14.0.2
pycodestyle==2.11.1
pycparser==2.23
pydantic==2.5.2
pydantic_core==2.14.5
pyflakes==3.2.0
Pygments==2.19.2
pymongo==4.6.1
pyparsing==3.2.5
pytest==7.4.4
pytest-cov==4.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==4.0.0
pytz==2023.3
pywinpty==3.0.2
PyYAML==6.0.3
pyzmq==27.1.0
qtconsole==5.7.0
QtPy==2.4.3
referencing==0.37.0
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==13.7.0
rpds-py==0.30.0
seaborn==0.13.0
Send2Trash==1.8.3
six==1.17.0
soupsieve==2.8
SQLAlchemy==2.0.25
stack-data==0.6.3
structlog==23.2.0
tenacity==9.1.2
terminado==0.18.1
tinycss2==1.4.0
tornado==6.5.2
tqdm==4.66.1
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
wcwidth==0.2.14
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
widgetsnbextension==4.0.15
yarl==1.22.0
